Fix avro parser tree Why? The last element in the key of the user is an attribute but was not treated as such What? Treat last element as attribute
Fix scalar case
Why? If we did had a scalar value we did not add anywhere the being mark or finish mark
to the shape builder
What? Added this to the increment method in the shape builder because this method knows
most of the context of the shape and can correctly handle the case; other ways could
have expliclitly kept the state of being an array element or not in the parsing and if
not within an array to add the begin and finish marks

Fix namespace resolution
Why? Namespace was added to the user keys -- even if the user has not provided these
What? Remove the default namespace that we use internally to simplify the parser trees
from the user-scope; however, if the user supplies a namespace we will check and
resolve it